backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
aiohttp==3.11.0
pytest==8.3.3
pytest-asyncio==0.24.0

# Note: xai-sdk is installed separately via: pip install git+https://github.com/xai-org/xai-sdk-python.git

